Extract `comm` module
Represents "communication" and includes `ingest`, `publish`, and `peer`
module. It contains shared items originally located in `main.rs`.

Having common items in `main.rs` makes it difficult to add additional
binaries bisides the main one.
