+++ src/comm/publish.rs
@@ -51,14 +51,16 @@ use tokio::{
 use tracing::{debug, error, info, warn};
 
 use self::implement::RequestStreamMessage;
+use crate::comm::{
+    IngestSensors, PcapSensors, StreamDirectChannels,
+    ingest::{NetworkKey, implement::EventFilter},
+    peer::{PeerIdents, Peers},
+};
 use crate::graphql::TIMESTAMP_SIZE;
-use crate::ingest::{NetworkKey, implement::EventFilter};
-use crate::peer::{PeerIdents, Peers};
 use crate::server::{
     Certs, config_client, config_server, extract_cert_from_conn, subject_from_cert_verbose,
 };
 use crate::storage::{Database, Direction, RawEventStore, StorageKey};
-use crate::{IngestSensors, PcapSensors, StreamDirectChannels};
 
 const PUBLISH_VERSION_REQ: &str = ">=0.23.0,<0.26.0";
