+++ src/graphql/status.rs
@@ -11,7 +11,7 @@ use crate::graphql::{StringNumberU32, StringNumberU64};
 use crate::settings::ConfigVisible;
 #[cfg(debug_assertions)]
 use crate::storage::Database;
-use crate::{peer::PeerIdentity, settings::Settings};
+use crate::{comm::peer::PeerIdentity, settings::Settings};
 
 const GRAPHQL_REBOOT_DELAY: u64 = 100;
 pub const CONFIG_PUBLISH_SRV_ADDR: &str = "publish_srv_addr";
